requirements.txt
absl-py==2.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bleach==6.3.0
blessed==1.20.0
blinker==1.9.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
comm==0.2.3
contexttimer==0.3.3
contourpy==1.3.1
cryptography==46.0.5
cursor==1.3.5
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
earth==0.0.1
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.12.0
fastjsonschema==2.21.2
filelock==3.25.2
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.55.3
fpdf==1.7.2
fqdn==1.5.1
fsspec==2026.3.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.80.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
humanfriendly==10.0
idna==3.11
imbalanced-learn==0.14.1
ipykernel==7.1.0
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
jiter==0.13.0
joblib==1.4.2
jpype1==1.6.0
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.15
keras==3.14.0
kiwisolver==1.4.7
langdetect==1.0.9
lark==1.3.1
libclang==18.1.1
lxml==5.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mars==0.0.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.4
mpmath==1.3.0
mvn==0.1.3
namex==0.1.0
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.2.0
openai==2.30.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.19.0
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
path==17.0.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.0.0
platformdirs==4.5.0
plotly==6.7.0
prometheus_client==0.23.1
prompt_toolkit==3.0.50
protobuf==6.33.2
psutil==7.1.2
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
pymongo==4.10.1
pyparsing==3.2.0
pypdfium2==5.5.0
PyQtWebEngine-Qt5==5.15.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2024.2
pyuca==1.2
pywinpty==3.0.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.4.4
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.28.0
safetensors==0.7.0
scikit-learn==1.6.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==5.4.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
srsly==2.5.1
stack-data==0.6.3
streamlit==1.56.0
sympy==1.14.0
tabula-py==2.10.0
tenacity==9.1.2
tensorflow==2.21.0
termcolor==3.3.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
torch==2.11.0
tornado==6.5.2
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.46.3
widgetsnbextension==4.0.14
wrapt==2.1.2
xgboost==3.2.0
